Change interface
Using typedefs was just as inflexible as using explicit
`ADD_XXX_OPTION`, since the user had to manually specify a type anyway,
but using imaginary constants that weren't actually defined anywhere,
since they were part of a macro expansion.

This reverts to an improved version of the previous interface style,
with macros inspired by the BSD implementation of tree(3):

https://svnweb.freebsd.org/base/head/sys/sys/tree.h?view=markup

---

- Added `OPT_TYPE` enum instead of separate `#defines`

- Removed `START_OPTION` and moved the arguments into `PROCESS_OPTION`
macro and the state-clearing to a new `OPT_CLEAR` macro.

- Additionally, identifiers have been updated so that all names not
explicitly declared by the user are prefixed with `OPT_`.

  - ADD_OPTION -> OPT_FLAG(), OPT_INT(), OPT_STR()

  - PROCESS_OPTION -> OPT_PROCESS